Días de la semana con platos sugeridos, mejoras en el disenio de colores material design 600, correcciones y mejoras en los permisos de los usuarios según su rol, estandarización de a región de búsqueda en los mapas implementado en la clase StringValue

+++ app/src/main/java/com/appcocha/llajtacomida/model/plate/PlateManagerModel.java
@@ -4,6 +4,7 @@ import android.net.Uri;
 import android.util.Log;
 import androidx.annotation.NonNull;
 import com.appcocha.llajtacomida.interfaces.PlateInterface;
+import com.appcocha.llajtacomida.presenter.tools.Validation;
 import com.google.android.gms.tasks.Continuation;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.OnFailureListener;
@@ -14,6 +15,7 @@ import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.storage.FirebaseStorage;
 import com.google.firebase.storage.StorageReference;
 import com.google.firebase.storage.UploadTask;
+
 import java.util.Objects;
 
 /**
@@ -49,20 +51,46 @@ public class PlateManagerModel implements PlateInterface.ModelPlateManager {
                 .child("plates")
                 .child(plateId)
                 .child(plateId+".jpg");
-        storageReference.delete().addOnSuccessListener(new OnSuccessListener<Void>() {
+//        storageReference.delete().addOnSuccessListener(new OnSuccessListener<Void>() {
+//            @Override
+//            public void onSuccess(Void aVoid) {
+//            databaseReference.removeValue().addOnCompleteListener(new OnCompleteListener<Void>() {
+//                @Override
+//                public void onComplete(@NonNull Task<Void> task) {
+//                    presenterPlateManager.isSuccess(task.isSuccessful());
+//                }
+//            });
+//            }
+//        }).addOnFailureListener(new OnFailureListener() {
+//            @Override
+//            public void onFailure(@NonNull Exception exception) {
+//                presenterPlateManager.isSuccess(false);
+//            }
+//        });
+
+//      Nueva forma: Tomando en cuenta las reglas de firebase (realtimedatabase)
+        databaseReference.removeValue().addOnCompleteListener(new OnCompleteListener<Void>() {
             @Override
-            public void onSuccess(Void aVoid) {
-            databaseReference.removeValue().addOnCompleteListener(new OnCompleteListener<Void>() {
-                @Override
-                public void onComplete(@NonNull Task<Void> task) {
-                    presenterPlateManager.isSuccess(task.isSuccessful());
+            public void onComplete(@NonNull Task<Void> task) {
+                if(task.isSuccessful()){
+                    storageReference.delete().addOnSuccessListener(new OnSuccessListener<Void>() {
+                        @Override
+                        public void onSuccess(Void unused) {
+//                            Log.d("xxxxx28", "---> " +true);
+                            presenterPlateManager.isSuccess(true);
+                        }
+                    }).addOnFailureListener(new OnFailureListener() {
+                        @Override
+                        public void onFailure(@NonNull Exception e) {
+//                            Log.d("xxxxx28", "---> " +false);
+                            Log.e("Error", e.getMessage());
+                            presenterPlateManager.isSuccess(false);
+                        }
+                    });
+                }else{
+//                    Log.d("xxxxx28", "---> " +false);
+                    presenterPlateManager.isSuccess(false);
                 }
-            });
-            }
-        }).addOnFailureListener(new OnFailureListener() {
-            @Override
-            public void onFailure(@NonNull Exception exception) {
-                presenterPlateManager.isSuccess(false);
             }
         });
     }
@@ -82,7 +110,6 @@ public class PlateManagerModel implements PlateInterface.ModelPlateManager {
                 .child("plates")
                 .child(plate.getId())
                 .child(plate.getId()+".jpg");
-
         if (thumb_byte != null) {
             uploadData(plate, thumb_byte);
         }else{
@@ -122,32 +149,69 @@ public class PlateManagerModel implements PlateInterface.ModelPlateManager {
      * @param thumb_byte
      */
     private void uploadData(final Plate plate, final byte [] thumb_byte){
-        UploadTask uploadTask = storageReference.putBytes(thumb_byte);
-        uploadTask.continueWithTask(new Continuation<UploadTask.TaskSnapshot, Task<Uri>>() {
+//        UploadTask uploadTask = storageReference.putBytes(thumb_byte);
+//        uploadTask.continueWithTask(new Continuation<UploadTask.TaskSnapshot, Task<Uri>>() {
+//            @Override
+//            public Task<Uri> then(@NonNull Task<UploadTask.TaskSnapshot> task) throws Exception {
+//                if(task.isSuccessful()){
+//                    return storageReference.getDownloadUrl();
+//                }else{
+//                    throw Objects.requireNonNull(task.getException());
+//                }
+//            }
+//        }).addOnCompleteListener(new OnCompleteListener<Uri>() {
+//            @Override
+//            public void onComplete(@NonNull Task<Uri> task) {
+//                if(task.isSuccessful()){ // Si se subió la imagen, procedemos al registro en la base de datos
+//                    plate.setUrl(task.getResult().toString());
+//                    databaseReference.setValue(plate).addOnCompleteListener(new OnCompleteListener<Void>() {
+//                        @Override
+//                        public void onComplete(@NonNull Task<Void> task) {
+//                            presenterPlateManager.isSuccess(true);
+//                        }
+//                    });
+//                }else{
+//                    presenterPlateManager.isSuccess(false);
+//                }
+//            }
+//        });
+
+// Nueva forma: Tomando en cuenta las reglas de firebase (realtimedatabase)
+        databaseReference.setValue(plate).addOnCompleteListener(new OnCompleteListener<Void>() {
             @Override
-            public Task<Uri> then(@NonNull Task<UploadTask.TaskSnapshot> task) throws Exception {
+            public void onComplete(@NonNull Task<Void> task) {
+                // cargar imagen
                 if(task.isSuccessful()){
-                    return storageReference.getDownloadUrl();
-                }else{
-                    throw Objects.requireNonNull(task.getException());
-                }
-            }
-        }).addOnCompleteListener(new OnCompleteListener<Uri>() {
-            @Override
-            public void onComplete(@NonNull Task<Uri> task) {
-                if(task.isSuccessful()){ // Si se subió la imagen, procedemos al registro en la base de datos
-                    plate.setUrl(task.getResult().toString());
-                    databaseReference.setValue(plate).addOnCompleteListener(new OnCompleteListener<Void>() {
+//                    Log.d("xxxxx28", "---> " +true);
+                    UploadTask uploadTask = storageReference.putBytes(thumb_byte);
+                    uploadTask.continueWithTask(new Continuation<UploadTask.TaskSnapshot, Task<Uri>>() {
                         @Override
-                        public void onComplete(@NonNull Task<Void> task) {
-                            presenterPlateManager.isSuccess(true);
+                        public Task<Uri> then(@NonNull Task<UploadTask.TaskSnapshot> task) throws Exception {
+                            if(task.isSuccessful()){
+                                return storageReference.getDownloadUrl();
+                            }else{
+                                throw Objects.requireNonNull(task.getException());
+                            }
+                        }
+                    }).addOnCompleteListener(new OnCompleteListener<Uri>() {
+                        @Override
+                        public void onComplete(@NonNull Task<Uri> task) {
+                            if(task.isSuccessful()){
+                                plate.setUrl(task.getResult().toString());
+                                update(plate, null); // agrega la url despues de que termina de cargar el archivo
+//                                presenterPlateManager.isSuccess(true);
+                            }else{
+                                presenterPlateManager.isSuccess(false);
+                            }
                         }
                     });
                 }else{
+//                    Log.d("xxxxx28", "---> " + false);
                     presenterPlateManager.isSuccess(false);
                 }
             }
         });
+
     }
 
     /**
@@ -156,28 +220,67 @@ public class PlateManagerModel implements PlateInterface.ModelPlateManager {
      * @param thumb_byte
      */
     private void updatePlate(final Plate plate, final byte [] thumb_byte){
-        UploadTask uploadTask = storageReference.putBytes(thumb_byte);
-        uploadTask.continueWithTask(new Continuation<UploadTask.TaskSnapshot, Task<Uri>>() {
+//        UploadTask uploadTask = storageReference.putBytes(thumb_byte);
+//        uploadTask.continueWithTask(new Continuation<UploadTask.TaskSnapshot, Task<Uri>>() {
+//            @Override
+//            public Task<Uri> then(@NonNull Task<UploadTask.TaskSnapshot> task) throws Exception {
+//                if(task.isSuccessful()){
+//                    return storageReference.getDownloadUrl();
+//                }else{
+//                    throw Objects.requireNonNull(task.getException());
+//                }
+//            }
+//        }).addOnCompleteListener(new OnCompleteListener<Uri>() {
+//            @Override
+//            public void onComplete(@NonNull Task<Uri> task) {
+//                if(task.isSuccessful()){ // Si se subió la imagen, procedemos al registro en la base de datos
+//                    plate.setUrl(task.getResult().toString());
+//                    databaseReference.updateChildren(plate.toMap()).addOnCompleteListener(new OnCompleteListener<Void>() {
+//                        @Override
+//                        public void onComplete(@NonNull Task<Void> task) {
+//                            presenterPlateManager.isSuccess(true);
+//                        }
+//                    });
+//                }else{
+//                    presenterPlateManager.isSuccess(false);
+//                }
+//            }
+//        });
+
+//  Nueva forma: Tomando en cuenta las reglas de firebase (realtimedatabase)
+        databaseReference.updateChildren(plate.toMap()).addOnCompleteListener(new OnCompleteListener<Void>() {
             @Override
-            public Task<Uri> then(@NonNull Task<UploadTask.TaskSnapshot> task) throws Exception {
+            public void onComplete(@NonNull Task<Void> task) {
                 if(task.isSuccessful()){
-                    return storageReference.getDownloadUrl();
-                }else{
-                    throw Objects.requireNonNull(task.getException());
-                }
-            }
-        }).addOnCompleteListener(new OnCompleteListener<Uri>() {
-            @Override
-            public void onComplete(@NonNull Task<Uri> task) {
-                if(task.isSuccessful()){ // Si se subió la imagen, procedemos al registro en la base de datos
-                    plate.setUrl(task.getResult().toString());
-                    databaseReference.updateChildren(plate.toMap()).addOnCompleteListener(new OnCompleteListener<Void>() {
+//                    Log.d("xxxxx28", "---> " +true);
+//                    presenterPlateManager.isSuccess(true);
+                    // subir imagen
+                    UploadTask uploadTask = storageReference.putBytes(thumb_byte);
+                    uploadTask.continueWithTask(new Continuation<UploadTask.TaskSnapshot, Task<Uri>>() {
                         @Override
-                        public void onComplete(@NonNull Task<Void> task) {
-                            presenterPlateManager.isSuccess(true);
+                        public Task<Uri> then(@NonNull Task<UploadTask.TaskSnapshot> task) throws Exception {
+                            if(task.isSuccessful()){
+                                return storageReference.getDownloadUrl();
+                            }else{
+                                throw Objects.requireNonNull(task.getException());
+                            }
+                        }
+                    }).addOnCompleteListener(new OnCompleteListener<Uri>() {
+                        @Override
+                        public void onComplete(@NonNull Task<Uri> task) {
+                            if(task.isSuccessful()){
+                                plate.setUrl(task.getResult().toString());
+                                databaseReference.updateChildren(plate.toMap()).addOnCompleteListener(new OnCompleteListener<Void>() {
+                                    @Override
+                                    public void onComplete(@NonNull Task<Void> task) {
+                                        presenterPlateManager.isSuccess(task.isSuccessful());
+                                    }
+                                });
+                            }
                         }
                     });
-                }else{
+                }else {
+//                    Log.d("xxxxx28", "---> " +true);
                     presenterPlateManager.isSuccess(false);
                 }
             }
