+++ app/src/main/java/com/appcocha/llajtacomida/model/restaurant/RestaurantManagerModel.java
@@ -43,22 +43,44 @@ public class RestaurantManagerModel implements RestaurantInterface.ModelRestaura
                 .child("restaurants")
                 .child(restaurantId)
                 .child(restaurantId+".jpg");
-        storageReference.delete().addOnSuccessListener(new OnSuccessListener<Void>() {
+//        storageReference.delete().addOnSuccessListener(new OnSuccessListener<Void>() {
+//            @Override
+//            public void onSuccess(Void aVoid) {
+//                databaseReference.removeValue().addOnCompleteListener(new OnCompleteListener<Void>() {
+//                    @Override
+//                    public void onComplete(@NonNull Task<Void> task) {
+//                        presenterRestaurantManager.isSuccess(task.isSuccessful());
+//                    }
+//                });
+////                Toast.makeText(context, "Se eliminó correctamente", Toast.LENGTH_SHORT).show();
+//            }
+//        }).addOnFailureListener(new OnFailureListener() {
+//            @Override
+//            public void onFailure(@NonNull Exception exception) {
+////                Toast.makeText(context, "Oh no, algo salió mal\n"+exception.getMessage(), Toast.LENGTH_SHORT).show();
+//                presenterRestaurantManager.isSuccess(false);
+//            }
+//        });
+
+
+        //      Nueva forma: Tomando en cuenta las reglas de firebase (realtimedatabase)
+        databaseReference.removeValue().addOnCompleteListener(new OnCompleteListener<Void>() {
             @Override
-            public void onSuccess(Void aVoid) {
-                databaseReference.removeValue().addOnCompleteListener(new OnCompleteListener<Void>() {
-                    @Override
-                    public void onComplete(@NonNull Task<Void> task) {
-                        presenterRestaurantManager.isSuccess(task.isSuccessful());
-                    }
-                });
-//                Toast.makeText(context, "Se eliminó correctamente", Toast.LENGTH_SHORT).show();
-            }
-        }).addOnFailureListener(new OnFailureListener() {
-            @Override
-            public void onFailure(@NonNull Exception exception) {
-//                Toast.makeText(context, "Oh no, algo salió mal\n"+exception.getMessage(), Toast.LENGTH_SHORT).show();
-                presenterRestaurantManager.isSuccess(false);
+            public void onComplete(@NonNull Task<Void> task) {
+                if(task.isSuccessful()){
+                    storageReference.delete().addOnSuccessListener(new OnSuccessListener<Void>() {
+                        @Override
+                        public void onSuccess(Void unused) {
+                            presenterRestaurantManager.isSuccess(true);
+                        }
+                    }).addOnFailureListener(new OnFailureListener() {
+                        @Override
+                        public void onFailure(@NonNull Exception e) {
+                            Log.e("Error", e.getMessage());
+                            presenterRestaurantManager.isSuccess(false);
+                        }
+                    });
+                }else presenterRestaurantManager.isSuccess(false);
             }
         });
     }
@@ -110,38 +132,67 @@ public class RestaurantManagerModel implements RestaurantInterface.ModelRestaura
                     presenterRestaurantManager.isSuccess(task.isSuccessful());
                 }
             });
-//            Toast.makeText(context, "Se actualizó correctamente", Toast.LENGTH_SHORT).show();
         }
     }
 
     // Almacena nuevo
     private void uploadData(final Restaurant restaurant, final byte [] thumb_byte){
-        UploadTask uploadTask = storageReference.putBytes(thumb_byte);
-        uploadTask.continueWithTask(new Continuation<UploadTask.TaskSnapshot, Task<Uri>>() {
+//        UploadTask uploadTask = storageReference.putBytes(thumb_byte);
+//        uploadTask.continueWithTask(new Continuation<UploadTask.TaskSnapshot, Task<Uri>>() {
+//            @Override
+//            public Task<Uri> then(@NonNull Task<UploadTask.TaskSnapshot> task) throws Exception {
+//                if(task.isSuccessful()){
+//                    return storageReference.getDownloadUrl();
+//                }else{
+//                    throw Objects.requireNonNull(task.getException());
+//                }
+//            }
+//        }).addOnCompleteListener(new OnCompleteListener<Uri>() {
+//            @Override
+//            public void onComplete(@NonNull Task<Uri> task) {
+//                if(task.isSuccessful()){ // Si se subió la imagen, procedemos al registro en la base de datos
+//                    restaurant.setUrl(task.getResult().toString());
+//                    databaseReference.setValue(restaurant).addOnCompleteListener(new OnCompleteListener<Void>() {
+//                        @Override
+//                        public void onComplete(@NonNull Task<Void> task) {
+//                            presenterRestaurantManager.isSuccess(true);
+//                        }
+//                    });
+//                }else{
+//                    presenterRestaurantManager.isSuccess(false);
+//                }
+//            }
+//        });
+
+        //      Nueva forma: Tomando en cuenta las reglas de firebase (realtimedatabase)
+        databaseReference.setValue(restaurant).addOnCompleteListener(new OnCompleteListener<Void>() {
             @Override
-            public Task<Uri> then(@NonNull Task<UploadTask.TaskSnapshot> task) throws Exception {
+            public void onComplete(@NonNull Task<Void> task) {
                 if(task.isSuccessful()){
-                    return storageReference.getDownloadUrl();
-                }else{
-                    throw Objects.requireNonNull(task.getException());
-                }
-            }
-        }).addOnCompleteListener(new OnCompleteListener<Uri>() {
-            @Override
-            public void onComplete(@NonNull Task<Uri> task) {
-                if(task.isSuccessful()){ // Si se subió la imagen, procedemos al registro en la base de datos
-//                    url = task.getResult();
-                    restaurant.setUrl(task.getResult().toString());
-                    databaseReference.setValue(restaurant).addOnCompleteListener(new OnCompleteListener<Void>() {
+                    UploadTask uploadTask = storageReference.putBytes(thumb_byte);
+                    uploadTask.continueWithTask(new Continuation<UploadTask.TaskSnapshot, Task<Uri>>() {
                         @Override
-                        public void onComplete(@NonNull Task<Void> task) {
-                            presenterRestaurantManager.isSuccess(true);
+                        public Task<Uri> then(@NonNull Task<UploadTask.TaskSnapshot> task) throws Exception {
+                            if(task.isSuccessful()){
+                                return storageReference.getDownloadUrl();
+                            }else{
+                                throw Objects.requireNonNull(task.getException());
+                            }
+                        }
+                    }).addOnCompleteListener(new OnCompleteListener<Uri>() {
+                        @Override
+                        public void onComplete(@NonNull Task<Uri> task) {
+                            if(task.isSuccessful()){ // Si se subió la imagen, procedemos al registro en la base de datos
+                                restaurant.setUrl(task.getResult().toString());
+//                                update(final Restaurant restaurant, final byte [] thumb_byte)
+                                update(restaurant, null); // vuelve a actualizar la url de la imagen
+                            }else{
+                                presenterRestaurantManager.isSuccess(false);
+                            }
                         }
                     });
-//                    Toast.makeText(context, "Se procesó correctamente", Toast.LENGTH_SHORT).show();
                 }else{
                     presenterRestaurantManager.isSuccess(false);
-//                    Toast.makeText(context, "No se pudo subir el registro solicitado, intentelo más tarde", Toast.LENGTH_SHORT).show();
                 }
             }
         });
@@ -153,31 +204,66 @@ public class RestaurantManagerModel implements RestaurantInterface.ModelRestaura
      * @param thumb_byte
      */
     private void updateRestaurant(final Restaurant restaurant, final byte [] thumb_byte){
-        UploadTask uploadTask = storageReference.putBytes(thumb_byte);
-        uploadTask.continueWithTask(new Continuation<UploadTask.TaskSnapshot, Task<Uri>>() {
+//        UploadTask uploadTask = storageReference.putBytes(thumb_byte);
+//        uploadTask.continueWithTask(new Continuation<UploadTask.TaskSnapshot, Task<Uri>>() {
+//            @Override
+//            public Task<Uri> then(@NonNull Task<UploadTask.TaskSnapshot> task) throws Exception {
+//                if(task.isSuccessful()){
+//                    return storageReference.getDownloadUrl();
+//                }else{
+//                    throw Objects.requireNonNull(task.getException());
+//                }
+//            }
+//        }).addOnCompleteListener(new OnCompleteListener<Uri>() {
+//            @Override
+//            public void onComplete(@NonNull Task<Uri> task) {
+//                if(task.isSuccessful()){ // Si se subió la imagen, procedemos al registro en la base de datos
+//                    //                url = task.getResult();
+//                    restaurant.setUrl(task.getResult().toString());
+//                    databaseReference.updateChildren(restaurant.toMap()).addOnCompleteListener(new OnCompleteListener<Void>() {
+//                        @Override
+//                        public void onComplete(@NonNull Task<Void> task) {
+//                            presenterRestaurantManager.isSuccess(true);
+//                        }
+//                    });
+//                }else{
+//                    presenterRestaurantManager.isSuccess(false);
+//                }
+//            }
+//        });
+
+
+        //      Nueva forma: Tomando en cuenta las reglas de firebase (realtimedatabase)
+        databaseReference.updateChildren(restaurant.toMap()).addOnCompleteListener(new OnCompleteListener<Void>() {
             @Override
-            public Task<Uri> then(@NonNull Task<UploadTask.TaskSnapshot> task) throws Exception {
+            public void onComplete(@NonNull Task<Void> task) {
                 if(task.isSuccessful()){
-                    return storageReference.getDownloadUrl();
-                }else{
-                    throw Objects.requireNonNull(task.getException());
-                }
-            }
-        }).addOnCompleteListener(new OnCompleteListener<Uri>() {
-            @Override
-            public void onComplete(@NonNull Task<Uri> task) {
-                if(task.isSuccessful()){ // Si se subió la imagen, procedemos al registro en la base de datos
-                    //                url = task.getResult();
-                    restaurant.setUrl(task.getResult().toString());
-                    databaseReference.updateChildren(restaurant.toMap()).addOnCompleteListener(new OnCompleteListener<Void>() {
+                    UploadTask uploadTask = storageReference.putBytes(thumb_byte);
+                    uploadTask.continueWithTask(new Continuation<UploadTask.TaskSnapshot, Task<Uri>>() {
                         @Override
-                        public void onComplete(@NonNull Task<Void> task) {
-                            presenterRestaurantManager.isSuccess(true);
+                        public Task<Uri> then(@NonNull Task<UploadTask.TaskSnapshot> task) throws Exception {
+                            if(task.isSuccessful()){
+                                return storageReference.getDownloadUrl();
+                            }else{
+                                throw Objects.requireNonNull(task.getException());
+                            }
+                        }
+                    }).addOnCompleteListener(new OnCompleteListener<Uri>() {
+                        @Override
+                        public void onComplete(@NonNull Task<Uri> task) {
+                            if(task.isSuccessful()){ // Si se subió la imagen, procedemos al registro en la base de datos
+                                restaurant.setUrl(task.getResult().toString());
+                                databaseReference.updateChildren(restaurant.toMap()).addOnCompleteListener(new OnCompleteListener<Void>() {
+                                    @Override
+                                    public void onComplete(@NonNull Task<Void> task) {
+                                        if(task.isSuccessful()) presenterRestaurantManager.isSuccess(true);
+                                        else presenterRestaurantManager.isSuccess(false);
+                                    }
+                                });
+                            }else presenterRestaurantManager.isSuccess(false);
                         }
                     });
-                }else{
-                    presenterRestaurantManager.isSuccess(false);
-                }
+                }else presenterRestaurantManager.isSuccess(false);
             }
         });
     }
